Port src/marks/dot from JS to TS (stage 1)
Inline the hand-written .d.ts types into dot.ts and delete the original
.js and .d.ts files. No behaviour changes: parameters, defaults, and
iteration order are preserved.

The shared modules (context, options, style, symbol, mark) still ship
.d.ts files that don't list every runtime export; @ts-expect-error
markers note where dot.ts depends on still-untyped helpers, to be
removed as those modules are ported.
